updates to diagram-rest module
