+++ diagram-rest/src/main/java/io/serverlessworkflow/reactive_api_rest/ServerlessRequest.java
@@ -15,7 +15,6 @@
  */
 package io.serverlessworkflow.reactive_api_rest;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.server.ServerRequest;
@@ -23,10 +22,7 @@ import org.springframework.web.reactive.function.server.ServerResponse;
 import reactor.core.publisher.Mono;
 
 @Component
-@RequiredArgsConstructor
 public class ServerlessRequest {
-
-  private final ServerlesRequestHelper helper;
 
   /**
    * Get the SVG diagram of a workflow from API Request
@@ -40,8 +36,8 @@ public class ServerlessRequest {
         .body(
             sRequest
                 .bodyToMono(String.class)
-                .flatMap(helper::getSvg)
-                .onErrorMap(e -> new ControllerErrorException(e.getMessage())),
+                .flatMap(ServerlesRequestHelper::getSvg)
+                .onErrorMap(e -> new IllegalArgumentException(e.getMessage())),
             ServerlessWorkFlowResponse.class);
   }
 }
