+++ diagram-rest/src/main/resources/templates/plantuml/custom-template.txt
