+++ diagram/pom.xml
@@ -13,7 +13,7 @@
     <name>Serverless Workflow :: Diagram</name>
     <version>${project.parent.version}</version>
     <packaging>jar</packaging>
-    <description>Java SDK for Serverless Workflow Specification</description>
+    <description>Diagram Generation</description>
 
     <dependencies>
         <dependency>
